squash! test/AspNetCore/HttpSysCommunicationListenerTest.cs
Address the following finding.

### ⚠️ Port generated as `fuzzy.Int32()` instead of `fuzzy.UInt16()` at [test/AspNetCore/HttpSysCommunicationListenerTest.cs#L85](test/AspNetCore/HttpSysCommunicationListenerTest.cs#L85)

*Reported by: gemini (💡), gpt (💡), opus (⚠️). Elevated to ⚠️ per multi-model consensus.*

Service Fabric endpoint ports are 16-bit unsigned. Sibling tests (`WebHostCommunicationListenerTest`, `GenericHostCommunicationListenerTest`) use `fuzzy.UInt16()`. `Int32()` can produce negative or out-of-range values misrepresenting the API. Per `fuzzy.instructions.md` constraint discipline, choose the type that matches the domain.

+++ test/AspNetCore/HttpSysCommunicationListenerTest.cs
@@ -89,7 +89,7 @@ public abstract class HttpSysCommunicationListenerTest
                 Name = fuzzy.String(),
                 Protocol = protocol == EndpointProtocol.Http ? EndpointProtocol.Https : EndpointProtocol.Http,
             };
-            other.Property<int>(nameof(EndpointResourceDescription.Port)).Set(fuzzy.Int32());
+            other.Property<int>(nameof(EndpointResourceDescription.Port)).Set(fuzzy.UInt16());
             context.CodePackageActivationContext.GetEndpoints().Add(other);
 
             var endpoint = new EndpointResourceDescription
